refactor: moved views/structs to separate files
